# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "GenerateRequest" : {
        "properties" : {
          "model" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "HandoffRequest" : {
        "properties" : {
          "target" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "target" ],
        "type" : "object"
      },
      "HandoffResponse" : {
        "properties" : {
          "command" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "documentPath" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "prompt" : {
            "type" : "string"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "taskId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "HandoffResult" : {
        "properties" : {
          "handoff" : {
            "$ref" : "#/components/schemas/HandoffResponse"
          },
          "prompt" : {
            "type" : "string"
          },
          "promptToClipboard" : {
            "type" : "boolean"
          },
          "task" : {
            "$ref" : "#/components/schemas/TaskResponse"
          }
        },
        "type" : "object"
      },
...
      "PhaseResponse" : {
        "properties" : {
          "approval" : {
            "enum" : [ "PENDING", "APPROVED", "CHANGES_REQUESTED" ],
            "type" : "string"
          },
          "approvalNote" : {
            "type" : "string"
          },
          "approvedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "documentPath" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "objective" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PLANNED", "IN_PROGRESS", "DONE" ],
            "type" : "string"
          },
          "tasks" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskResponse"
            },
            "type" : "array"
          },
          "title" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PlanResponse" : {
        "properties" : {
          "phases" : {
            "items" : {
              "$ref" : "#/components/schemas/PhaseResponse"
            },
            "type" : "array"
          },
          "project" : {
            "$ref" : "#/components/schemas/ProjectResponse"
          },
          "runs" : {
            "items" : {
              "$ref" : "#/components/schemas/PlanRunResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PlanRunResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "failureDetail" : {
            "type" : "string"
          },
          "finishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "inputTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "output" : {
            "type" : "string"
          },
          "outputTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "phases" : {
            "format" : "int32",
            "type" : "integer"
          },
          "progress" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "requestedModel" : {
            "type" : "string"
          },
          "servedModel" : {
            "type" : "string"
          },
          "source" : {
            "enum" : [ "ENGINE", "IMPORT" ],
            "type" : "string"
          },
          "status" : {
            "enum" : [ "RUNNING", "SUCCEEDED", "FAILED" ],
            "type" : "string"
          },
          "tasks" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ProjectResponse" : {
        "properties" : {
          "autonomyLevel" : {
            "enum" : [ "GUIDED", "SUPERVISED", "DELEGATED", "FINAL_REVIEW" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "planStatus" : {
            "enum" : [ "NONE", "DRAFT", "APPROVED" ],
            "type" : "string"
          },
          "projectType" : {
            "enum" : [ "WEB_APP", "BACKEND", "MOBILE", "DESKTOP", "AI_ML", "GAME", "THREE_D", "DATA", "AUTOMATION", "API", "FULL_STACK", "OTHER" ],
            "type" : "string"
          },
          "stack" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "ARCHIVED" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "workspacePath" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReviewRequest" : {
        "properties" : {
          "handoffId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "note" : {
            "maxLength" : 4000,
            "minLength" : 0,
            "type" : "string"
          },
          "runId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "verdict" : {
            "enum" : [ "ACCEPTED", "CHANGES_REQUESTED" ],
            "type" : "string"
          }
        },
        "required" : [ "verdict" ],
        "type" : "object"
      },
...
      "TaskCreateRequest" : {
        "properties" : {
          "agentId" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          },
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "priority" : {
            "minLength" : 1,
            "type" : "string"
          },
          "projectId" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "minLength" : 1,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "priority", "status", "title" ],
        "type" : "object"
      },
...
      "TaskResponse" : {
        "properties" : {
          "agentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "documentPath" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "phaseId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH" ],
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "OPEN", "IN_PROGRESS", "DONE" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Every route needs `Authorization: Bearer <token>` (ADR-013). Every mutation of an existing resource needs `If-Match` with the ETag last read (ADR-009): absent is 428, stale is 412. Every error is `application/problem+json` with a stable `type` = `urn:ai-company-os:problem:<slug>` (ADR-007); branch on `type`, not on `title`.",
    "title" : "AI Company OS — control plane",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/phases/{phaseId}/approve" : {
      "post" : {
        "operationId" : "approvePhase",
        "parameters" : [ {
          "in" : "path",
          "name" : "phaseId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PhaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/phases/{phaseId}/request-changes" : {
      "post" : {
        "operationId" : "requestChanges",
        "parameters" : [ {
          "in" : "path",
          "name" : "phaseId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PhaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/plan-runs/{runId}" : {
      "get" : {
        "operationId" : "run_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "runId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanRunResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
...
    "/api/projects/{id}/plan" : {
      "get" : {
        "operationId" : "plan",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/projects/{id}/plan/approve" : {
      "post" : {
        "operationId" : "approve",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "approveAllPhases",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/projects/{id}/plan/generate" : {
      "post" : {
        "operationId" : "generate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanRunResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/projects/{id}/plan/handoff" : {
      "get" : {
        "operationId" : "handoff_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
    "/api/projects/{id}/plan/import" : {
      "post" : {
        "operationId" : "importPlan",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanRunResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "plan-controller" ]
      }
    },
...
    "/api/tasks" : {
      "get" : {
        "operationId" : "getTasks",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "OPEN", "IN_PROGRESS", "DONE" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TaskResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      },
      "post" : {
        "operationId" : "createTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
...
    "/api/tasks/{id}/handoffs" : {
      "get" : {
        "operationId" : "handoffs",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/HandoffResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "orchestrator-controller" ]
      },
      "post" : {
        "operationId" : "handoff",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HandoffResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "orchestrator-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ]
}